+++ src/main/java/kitchenpos/table/domain/OrderTables.java
@@ -1,16 +1,10 @@
 package kitchenpos.table.domain;
 
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.table.domain.OrderTable;
-import kitchenpos.table.domain.OrderTableEntity;
-import kitchenpos.table.dto.OrderTableRequest;
 import kitchenpos.tablegroup.domain.TableGroupEntity;
-import org.springframework.util.CollectionUtils;
 
 import javax.persistence.CascadeType;
 import javax.persistence.Embeddable;
 import javax.persistence.OneToMany;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
